crud-admin/requirements.txt
django==4.2.5
django-rest-framework==0.1.0
django-cors-headers==4.2.0
psycopg2-binary==2.9.7
django-environ==0.11.2
django-admin-interface==0.26.1
django-requests==0.0.2
django-q==1.3.9
jsonfield==3.1.0
fastapi==0.109.0
elasticsearch==8.12.0

boto3==1.18.64
botocore==1.21.64
django-storages==1.11.1
gitpython
uvicorn 
asgiref
django-import-export
django-silk
django-redis
pandas
openpyxl
